requirements.txt
amqp==2.4.2
Babel==2.6.0
billiard==3.5.0.5
celery==4.2.2
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-cleanup==3.2.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-js-asset==1.2.2
django-mptt==0.9.1
django-phonenumber-field==2.2.0
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
djangorestframework==3.9.2
djangorestframework-recursive==0.1.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
kombu==4.3.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
phonenumbers==8.10.8
Pillow==5.4.1
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
redis==3.2.1
requests==2.21.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
